+++ DLSlideView/DLSlideView.h
@@ -11,23 +11,23 @@
 @class DLSlideView;
 
 @protocol DLSlideViewDataSource <NSObject>
-- (int)numberOfControllersInDLSlideView:(DLSlideView *)sender;
-- (UIViewController *)DLSlideView:(DLSlideView *)sender controllerAt:(int)index;
+- (NSInteger)numberOfControllersInDLSlideView:(DLSlideView *)sender;
+- (UIViewController *)DLSlideView:(DLSlideView *)sender controllerAt:(NSInteger)index;
 @end
 
 @protocol DLSlideViewDelegate <NSObject>
 @optional
-- (void)DLSlideView:(DLSlideView *)slide switchingFrom:(int)oldIndex to:(int)toIndex percent:(float)percent;
-- (void)DLSlideView:(DLSlideView *)slide didSwitchTo:(int)index;
-- (void)DLSlideView:(DLSlideView *)slide switchCanceled:(int)oldIndex;
+- (void)DLSlideView:(DLSlideView *)slide switchingFrom:(NSInteger)oldIndex to:(NSInteger)toIndex percent:(float)percent;
+- (void)DLSlideView:(DLSlideView *)slide didSwitchTo:(NSInteger)index;
+- (void)DLSlideView:(DLSlideView *)slide switchCanceled:(NSInteger)oldIndex;
 @end
 
 @interface DLSlideView : UIView
 //@property(nonatomic, strong) NSArray *viewControllers;
-@property(nonatomic, assign) int selectedIndex;
+@property(nonatomic, assign) NSInteger selectedIndex;
 @property(nonatomic, weak) UIViewController *baseViewController;
 @property(nonatomic, weak) id<DLSlideViewDelegate>delegate;
 @property(nonatomic, weak) id<DLSlideViewDataSource>dataSource;
-- (void)switchTo:(int)index;
+- (void)switchTo:(NSInteger)index;
 
 @end
