+++ DLSlideView/DLTabbarView/DLSlideTabbarProtocol.h
@@ -9,13 +9,13 @@
 #import <Foundation/Foundation.h>
 
 @protocol DLSlideTabbarDelegate <NSObject>
-- (void)DLSlideTabbar:(id)sender selectAt:(int)index;
+- (void)DLSlideTabbar:(id)sender selectAt:(NSInteger)index;
 @end
 
 @protocol DLSlideTabbarProtocol <NSObject>
-@property(nonatomic, assign) int selectedIndex;
-@property(nonatomic, readonly) int tabbarCount;
+@property(nonatomic, assign) NSInteger selectedIndex;
+@property(nonatomic, readonly) NSInteger tabbarCount;
 @property(nonatomic, weak) id<DLSlideTabbarDelegate> delegate;
-- (void)switchingFrom:(int)fromIndex to:(int)toIndex percent:(float)percent;
+- (void)switchingFrom:(NSInteger)fromIndex to:(NSInteger)toIndex percent:(float)percent;
 
 @end
